fix: add patient
1. fixed some logical errors

2. added sending the generated password to the patient by email

3. also fixed some minor logical errors in other moduels

+++ src/main/java/com/na7ki/backend/domain/user/model/create_patient/UserDetailsData.java
@@ -24,7 +24,7 @@ public record UserDetailsData(
         @Min(value = 0, message = "age must be between 0 and 20")
         @Max(value = 20, message = "age must be between 0 and 20")
         @Digits(integer = 2, fraction = 0, message = "age should not contain fractions")
-        Short age
+        Byte age
 
 ) {
 }
